feat: Use 'pip-secure-install' recommendations
Use Brett Cannon's recommendations for pip-secure-install to ensure
that the environment is fully reproducible and can be installed as
securely as possible. This works well with pip-tools.

c.f.:
* https://www.python.org/dev/peps/pep-0665/#secure-by-design
* https://github.com/brettcannon/pip-secure-install
* https://twitter.com/brettsky/status/1486137764315688961

The options used to follow pip-secure-install recommendations are:

* A requirements file must be specified to make sure all dependencies
are known statically for auditing purposes (--requirement).
* No dependency resolution is done to make sure the requirements file
is complete (--no-deps).
* All requirements must have a hash provided to make sure the files have
not been tampered with (--require-hashes).
* Only wheels are allowed to have reproducible installs (--only-binary :all:).

In the case of the libraries termcolor and jax, which only ship sdists, the
--only-binary option will fail and so the --no-binary option needs to be given
explicitly for them.